PDFBOX-2147: Removed unneccesary unboxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1609404 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdfparser/PDFXrefStreamParser.java
@@ -145,7 +145,7 @@ public class PDFXrefStreamParser extends BaseParser
                         {
                             genNum += (currLine[i + w0 + w1] & 0x00ff) << ((w2 - i - 1) * 8);
                         }
-                        COSObjectKey objKey = new COSObjectKey(objID.intValue(), genNum);
+                        COSObjectKey objKey = new COSObjectKey(objID, genNum);
                         xrefTrailerResolver.setXRef(objKey, offset);
                         break;
                     case 2:
@@ -168,7 +168,7 @@ public class PDFXrefStreamParser extends BaseParser
                         {
                             objstmObjNr += (currLine[i + w0] & 0x00ff) << ((w1 - i - 1) * 8);
                         }    
-                        objKey = new COSObjectKey( objID.intValue(), 0 );
+                        objKey = new COSObjectKey( objID, 0 );
                         xrefTrailerResolver.setXRef( objKey, -objstmObjNr );
                         break;
                     default:
