+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/common/function/type4/ExecutionContext.java
@@ -73,7 +73,7 @@ public class ExecutionContext
      */
     public int popInt()
     {
-        return ((Integer)stack.pop()).intValue();
+        return ((Integer)stack.pop());
     }
 
     /**
