+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/interactive/form/PDField.java
@@ -317,7 +317,7 @@ public abstract class PDField implements COSObjectable
         Integer ff = fdfField.getFieldFlags();
         if (ff != null)
         {
-            setFieldFlags(ff.intValue());
+            setFieldFlags(ff);
         }
         else
         {
@@ -326,7 +326,7 @@ public abstract class PDField implements COSObjectable
 
             if (setFf != null)
             {
-                int setFfInt = setFf.intValue();
+                int setFfInt = setFf;
                 fieldFlags = fieldFlags | setFfInt;
                 setFieldFlags(fieldFlags);
             }
@@ -342,7 +342,7 @@ public abstract class PDField implements COSObjectable
                 // clrFf = 1101
                 // clrFfValue = 0010;
                 // newValue = 1011 & 0010 which is 0010
-                int clrFfValue = clrFf.intValue();
+                int clrFfValue = clrFf;
                 clrFfValue ^= 0xFFFFFFFF;
                 fieldFlags = fieldFlags & clrFfValue;
                 setFieldFlags(fieldFlags);
@@ -356,7 +356,7 @@ public abstract class PDField implements COSObjectable
             Integer f = fdfField.getWidgetFieldFlags();
             if (f != null && widget != null)
             {
-                widget.setAnnotationFlags(f.intValue());
+                widget.setAnnotationFlags(f);
             }
             else
             {
@@ -364,7 +364,7 @@ public abstract class PDField implements COSObjectable
                 Integer setF = fdfField.getSetWidgetFieldFlags();
                 if (setF != null)
                 {
-                    annotFlags = annotFlags | setF.intValue();
+                    annotFlags = annotFlags | setF;
                     widget.setAnnotationFlags(annotFlags);
                 }
 
@@ -379,7 +379,7 @@ public abstract class PDField implements COSObjectable
                     // clrF = 1101
                     // clrFValue = 0010;
                     // newValue = 1011 & 0010 which is 0010
-                    int clrFValue = clrF.intValue();
+                    int clrFValue = clrF;
                     clrFValue ^= 0xFFFFFFFFL;
                     annotFlags = annotFlags & clrFValue;
                     widget.setAnnotationFlags(annotFlags);
